+++ web/package.json
@@ -30,7 +30,6 @@
     "react-dom": "^16.11.0",
     "react-google-login": "^5.0.7",
     "react-kakao-login": "^1.2.0",
-    "react-naver-login": "^0.1.3-alpha1",
     "react-router-dom": "^5.1.2",
     "short-uuid": "^3.1.1",
     "styled-components": "^4.4.1",
