+++ web/src/components/base/index.js
@@ -1,2 +1,3 @@
 export { default as Layout } from './Layout';
 export { default as ChannelProvider } from './ChannelProvider';
+export { default as NaverLogin } from './NaverLogin';
