--- web/src/pages/Login/index.js
@@ -1,1 +1,0 @@
-export { default } from './Login';
