+++ web/src/pages/index.js
@@ -1,5 +1,4 @@
 export { default as Main } from './Main';
 export { default as Channel } from './Channel';
 export { default as MyPage } from './MyPage';
-export { default as Login } from './Login';
 export { default as PrivacyPolicy } from './PrivacyPolicy';
